requirements.txt
# Core dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Deep SORT
deep-sort-realtime>=1.3.2

# Optional: for better Re-ID
# torch>=2.0.0
# torchvision>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils
tqdm>=4.65.0
pillow>=10.0.0